Working on sory

+++ src/Classes/Bond.java
@@ -5,7 +5,7 @@ public class Bond extends Assets{
     //Bonds do not have randomness to them, but rather a flat rate.
     private double rate;
     
-    public Bond(int v, String n, String c, double r){
+    public Bond(double v, String n, String c, double r){
         super(v,n,c,1,1);
         type = "BOND"; 
         rate = r;
